# examples/nest-market/openapi.json
{
  "openapi": "3.0.0",
  "paths": {
...
    "/api/v1/investors/individuals": {
      "post": {
        "operationId": "createIndividualInvestor",
        "parameters": [],
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/CreateIndividualInvestorDto"
              }
            }
          }
        },
        "responses": {
          "201": {
            "description": "",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/IndividualInvestorDto"
                }
              }
            }
          },
          "400": {
            "description": "",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ProblemDto"
                }
              }
            }
          },
          "401": {
            "description": "",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ProblemDto"
                }
              }
            }
          },
          "404": {
            "description": "",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ProblemDto"
                }
              }
            }
          }
        },
        "tags": [
          "investors"
        ],
        "security": [
          {
            "bearer": []
          }
        ]
      }
    },
    "/api/v1/investors/companies": {
      "post": {
        "operationId": "createCompanyInvestor",
        "parameters": [],
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/CreateCompanyInvestorDto"
              }
            }
          }
        },
        "responses": {
          "201": {
            "description": "",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/CompanyInvestorDto"
                }
              }
            }
          },
          "400": {
            "description": "",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ProblemDto"
                }
              }
            }
          },
          "401": {
            "description": "",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ProblemDto"
                }
              }
            }
          },
          "404": {
            "description": "",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ProblemDto"
                }
              }
            }
          }
        },
        "tags": [
          "investors"
        ],
        "security": [
          {
            "bearer": []
          }
        ]
      }
    },
    "/api/v1/investors/{investorId}": {
      "get": {
        "operationId": "getInvestor",
        "summary": "Read an individual or company investor, discriminated by kind",
        "parameters": [
          {
            "name": "investorId",
            "required": true,
            "in": "path",
            "schema": {
              "format": "uuid",
              "type": "string"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "",
            "content": {
              "application/json": {
                "schema": {
                  "oneOf": [
                    {
                      "$ref": "#/components/schemas/IndividualInvestorDto"
                    },
                    {
                      "$ref": "#/components/schemas/CompanyInvestorDto"
                    }
                  ],
                  "discriminator": {
                    "propertyName": "kind",
                    "mapping": {
                      "individual": "#/components/schemas/IndividualInvestorDto",
                      "company": "#/components/schemas/CompanyInvestorDto"
                    }
                  }
                }
              }
            }
          },
          "400": {
            "description": "",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ProblemDto"
                }
              }
            }
          },
          "401": {
            "description": "",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ProblemDto"
                }
              }
            }
          },
          "404": {
            "description": "",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ProblemDto"
                }
              }
            }
          }
        },
        "tags": [
          "investors"
        ],
        "security": [
          {
            "bearer": []
          }
        ]
      }
    },
...
    "/api/v1/offerings/{offeringId}/documents": {
      "post": {
        "operationId": "uploadDocument",
        "summary": "Upload an offering document with its category and note",
        "parameters": [
          {
            "name": "offeringId",
            "required": true,
            "in": "path",
            "schema": {
              "format": "uuid",
              "type": "string"
            }
          }
        ],
        "requestBody": {
          "required": true,
          "content": {
            "multipart/form-data": {
              "schema": {
                "$ref": "#/components/schemas/UploadDocumentDto"
              }
            }
          }
        },
        "responses": {
          "201": {
            "description": "",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/DocumentDto"
                }
              }
            }
          },
          "400": {
            "description": "",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ProblemDto"
                }
              }
            }
          },
          "401": {
            "description": "",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ProblemDto"
                }
              }
            }
          },
          "404": {
            "description": "",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ProblemDto"
                }
              }
            }
          },
          "413": {
            "description": "",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ProblemDto"
                }
              }
            }
          }
        },
        "tags": [
          "documents"
        ],
        "security": [
          {
            "bearer": []
          }
        ]
      }
    },
    "/api/v1/documents/{documentId}": {
      "get": {
        "operationId": "getDocument",
        "parameters": [
          {
            "name": "documentId",
            "required": true,
            "in": "path",
            "schema": {
              "format": "uuid",
              "type": "string"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/DocumentDto"
                }
              }
            }
          },
          "400": {
            "description": "",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ProblemDto"
                }
              }
            }
          },
          "401": {
            "description": "",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ProblemDto"
                }
              }
            }
          },
          "404": {
            "description": "",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ProblemDto"
                }
              }
            }
          }
        },
        "tags": [
          "documents"
        ],
        "security": [
          {
            "bearer": []
          }
        ]
      },
      "delete": {
        "operationId": "deleteDocument",
        "parameters": [
          {
            "name": "documentId",
            "required": true,
            "in": "path",
            "schema": {
              "format": "uuid",
              "type": "string"
            }
          }
        ],
        "responses": {
          "204": {
            "description": ""
          },
          "400": {
            "description": "",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ProblemDto"
                }
              }
            }
          },
          "401": {
            "description": "",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ProblemDto"
                }
              }
            }
          },
          "404": {
            "description": "",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ProblemDto"
                }
              }
            }
          }
        },
        "tags": [
          "documents"
        ],
        "security": [
          {
            "bearer": []
          }
        ]
      }
    },
    "/api/v1/documents/{documentId}/content": {
      "get": {
        "operationId": "downloadDocument",
        "parameters": [
          {
            "name": "documentId",
            "required": true,
            "in": "path",
            "schema": {
              "format": "uuid",
              "type": "string"
            }
          }
        ],
        "responses": {
          "200": {
            "content": {
              "application/octet-stream": {
                "schema": {
                  "type": "string",
                  "format": "binary"
                }
              }
            },
            "headers": {
              "Content-Disposition": {
                "schema": {
                  "type": "string"
                },
                "description": "Attachment filename"
              }
            },
            "description": ""
          },
          "400": {
            "description": "",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ProblemDto"
                }
              }
            }
          },
          "401": {
            "description": "",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ProblemDto"
                }
              }
            }
          },
          "404": {
            "description": "",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ProblemDto"
                }
              }
            }
          }
        },
        "tags": [
          "documents"
        ],
        "security": [
          {
            "bearer": []
          }
        ]
      }
    }
  },
  "info": {
    "title": "Harbor investment platform",
    "description": "A runnable NestJS example: offerings, investors, subscriptions and documents.",
    "version": "1.0.0",
    "contact": {}
  },
...
  "servers": [
    {
      "url": "http://127.0.0.1:3100",
      "description": "Local example server"
    }
  ],
  "components": {
    "securitySchemes": {
      "bearer": {
        "scheme": "bearer",
        "bearerFormat": "JWT",
        "type": "http"
      }
    },
    "schemas": {
      "ProblemDto": {
        "type": "object",
        "properties": {
          "statusCode": {
            "type": "integer",
            "example": 400
          },
          "code": {
            "type": "string",
            "example": "BAD_REQUEST"
          },
          "message": {
            "type": "string",
            "example": "Request validation failed"
          },
          "details": {
            "type": "array",
            "items": {
              "type": "string"
            }
          }
        },
        "required": [
          "statusCode",
          "code",
          "message"
        ]
      },
...
      "CreateIndividualInvestorDto": {
        "type": "object",
        "properties": {
          "displayName": {
            "type": "string",
            "minLength": 2,
            "example": "Alex Morgan"
          },
          "email": {
            "type": "string",
            "format": "email",
            "example": "alex@example.test"
          },
          "country": {
            "type": "string",
            "minLength": 2,
            "maxLength": 2,
            "example": "FR"
          },
          "onboardingNote": {
            "type": "string",
            "writeOnly": true,
            "description": "An onboarding note accepted on creation and omitted from public responses."
          },
          "kind": {
            "type": "string",
            "enum": [
              "individual"
            ]
          }
        },
        "required": [
          "displayName",
          "email",
          "country",
          "kind"
        ]
      },
      "IndividualInvestorDto": {
        "type": "object",
        "properties": {
          "displayName": {
            "type": "string",
            "minLength": 2,
            "example": "Alex Morgan"
          },
          "email": {
            "type": "string",
            "format": "email",
            "example": "alex@example.test"
          },
          "country": {
            "type": "string",
            "minLength": 2,
            "maxLength": 2,
            "example": "FR"
          },
          "kind": {
            "type": "string",
            "enum": [
              "individual"
            ]
          },
          "id": {
            "type": "string",
            "format": "uuid",
            "readOnly": true
          }
        },
        "required": [
          "displayName",
          "email",
          "country",
          "kind",
          "id"
        ]
      },
      "CreateCompanyInvestorDto": {
        "type": "object",
        "properties": {
          "displayName": {
            "type": "string",
            "minLength": 2,
            "example": "Alex Morgan"
          },
          "email": {
            "type": "string",
            "format": "email",
            "example": "alex@example.test"
          },
          "country": {
            "type": "string",
            "minLength": 2,
            "maxLength": 2,
            "example": "FR"
          },
          "onboardingNote": {
            "type": "string",
            "writeOnly": true,
            "description": "An onboarding note accepted on creation and omitted from public responses."
          },
          "kind": {
            "type": "string",
            "enum": [
              "company"
            ]
          },
          "registrationNumber": {
            "type": "string",
            "minLength": 3,
            "example": "REG-123456"
          }
        },
        "required": [
          "displayName",
          "email",
          "country",
          "kind",
          "registrationNumber"
        ]
      },
      "CompanyInvestorDto": {
        "type": "object",
        "properties": {
          "displayName": {
            "type": "string",
            "minLength": 2,
            "example": "Alex Morgan"
          },
          "email": {
            "type": "string",
            "format": "email",
            "example": "alex@example.test"
          },
          "country": {
            "type": "string",
            "minLength": 2,
            "maxLength": 2,
            "example": "FR"
          },
          "kind": {
            "type": "string",
            "enum": [
              "company"
            ]
          },
          "registrationNumber": {
            "type": "string",
            "minLength": 3,
            "example": "REG-123456"
          },
          "id": {
            "type": "string",
            "format": "uuid",
            "readOnly": true
          }
        },
        "required": [
          "displayName",
          "email",
          "country",
          "kind",
          "registrationNumber",
          "id"
        ]
      },
...
      "DocumentCategory": {
        "type": "string",
        "enum": [
          "terms",
          "prospectus",
          "other"
        ]
      },
      "UploadDocumentDto": {
        "type": "object",
        "properties": {
          "category": {
            "$ref": "#/components/schemas/DocumentCategory"
          },
          "note": {
            "type": "string"
          },
          "file": {
            "type": "string",
            "format": "binary"
          }
        },
        "required": [
          "category",
          "file"
        ]
      },
      "DocumentDto": {
        "type": "object",
        "properties": {
          "category": {
            "$ref": "#/components/schemas/DocumentCategory"
          },
          "note": {
            "type": "string"
          },
          "id": {
            "type": "string",
            "format": "uuid",
            "readOnly": true
          },
          "offeringId": {
            "type": "string",
            "format": "uuid",
            "readOnly": true
          },
          "filename": {
            "type": "string"
          },
          "size": {
            "type": "integer",
            "minimum": 0
          }
        },
        "required": [
          "category",
          "id",
          "offeringId",
          "filename",
          "size"
        ]
      }
    }
  }
}